requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.2.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==3.3
matplotlib-inline==0.1.7
mechanize==0.2.5
nest-asyncio==1.6.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.1
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
wcwidth==0.2.13
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.7_1/libexec/wheel-0.44.0-py3-none-any.whl#sha256=c01326144883c50263d6967d1145640f3bbc8dfab32c2cbe817ef61d7a810219
